Fix Styles If Already Assigned
The fallback element will always be set as a block element as it currently is. I needed the ability to set it as hidden on load, so I set the script to only set to block if it doesn't exist. The audio height also was set to always be 26px high. People may want to set this also, so I put through the getDimension function.